backend/rpa-bots/requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0